+++ inc/class.editors-add-users.php
@@ -126,6 +126,7 @@ class EditorsAddUsers extends Singleton {
     $user = wp_get_current_user();
     if( !current_user_can('administrator') ) {
       unset( $roles['administrator'] );
+      unset( $roles['editor_can_update'] );
     }
     return $roles;
   }